requirements.txt
a2wsgi==1.10.8
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
attr==0.3.2
attrs==25.3.0
babel==2.17.0
billiard==4.2.1
boto3==1.38.13
botocore==1.38.13
Brotli==1.1.0
brotlicffi==1.1.0.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configparser==7.2.0
cryptography==44.0.3
curio==1.6
Cython==3.1.0
decorator==5.2.1
dnspython==2.7.0
docopt==0.6.2
docutils==0.21.2
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fanstatic==1.5
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flower==2.0.1
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
HTMLParser==0.0.2
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
hypothesis==6.131.15
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==8.36.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
js==1.0
jupyterlab_widgets==3.0.15
keyring==25.6.0
kombu==5.5.3
legacy-cgi==2.6.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.1.2
multipart==1.2.1
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pipreqs==0.4.13
pluggy==1.5.0
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.5
roman-numerals-py==3.1.0
rsa==4.9.1
s3transfer==0.12.0
setuptools==80.4.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.46.2
textblob==0.19.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
urllib3-secure-extra==0.1.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
WebOb==1.8.9
websockets==15.0.1
widgetsnbextension==4.0.14
wsproto==1.2.0
yarg==0.1.10
zipp==3.21.0
zstandard==0.23.0

ollama~=0.4.8